Some fixes but mostly lots of cleanup in code and deletion of unused files
