--- Kristianstad/Source/Kristianstad/Business/Compare/OrganisationalUnit.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace Kristianstad.Business.Compare
-{
-    public class OrganisationalUnitModel
-    {
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-    }
-}
