--- Kristianstad/Source/Kristianstad/Controllers/Compare/API/CompareAPIController.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Web.Http;
-
-namespace Kristianstad.Controllers.Compare.API
-{
-    public class CompareAPIController : ApiController
-    {
-    }
-}
