+++ Kristianstad/Source/Kristianstad/ViewModels/Compare/OrganisationalUnitPageModel.cs
@@ -19,19 +19,10 @@ namespace Kristianstad.ViewModels.Compare
 
         public OrganisationalUnitPage CurrentPage { get; set; }
 
-        [UIHint(UIHint.Block)]
-        public OrganisationalUnitBlock OrganisationalUnitBlock { get; set; }
+        // [UIHint(UIHint.Block)]
+        // public OrganisationalUnitBlock OrganisationalUnitBlock { get; set; }
 
-        public IEnumerable<CategoryItemModel> Categories { get; set; }
+        // public IEnumerable<CategoryItemModel> Categories { get; set; }
      
-        public string PreviewText { get; set; }
-        public DateTime StartPublish { get; set; }
-        public XhtmlString MainBody { get; set; }
-
-        //public bool ShowPublishDate { get; set; }
-        //public bool ShowIntroduction { get; set; }
-
-        public CategoryList Category { get; set; }
-
     }
 }
