+++ Kristianstad/Source/Kristianstad/Views/Category/Full.cshtml
@@ -9,7 +9,7 @@
     <div @Html.EditAttributes("OrganisationalUnitList")>
 
         @Html.PropertyFor(x => x.CurrentPage.OrganisationalUnitList)
-        
+
     </div>
 </div>
