+++ Kristianstad/Source/Kristianstad/Views/Category/Index.cshtml
@@ -4,11 +4,6 @@
 
 @model Kristianstad.ViewModels.Compare.CategoryPageModel
 
-@{
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-
 <div class="column small-12 medium-12 large-8 content-page content">
     @{
         Html.RenderAction("Breadcrumbs", "MenuKr");
