--- Kristianstad/Source/Kristianstad/Views/CategoryPage/Index.cshtml
@@ -1,160 +1,0 @@
-﻿@using Kristianstad.Models.Pages
-@using EPiServer.Editor
-@using EPiServer.Security
-
-@model EPiCore.ViewModels.Pages.PageViewModel<CategoryPage>
-
-@{
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-
-<div class="column small-12 medium-12 large-8 content-page content">
-    @{
-        Html.RenderAction("Breadcrumbs", "MenuKr");
-    }
-
-    <h1 @Html.EditAttributes(x => x.CurrentPage.PageName)>@Model.CurrentPage.PageName</h1>
-    <p>
-        <div @Html.EditAttributes("PageStartPublish")>
-            @Html.PropertyFor(x => x.CurrentPage.StartPublish, new EPiServer.Web.Mvc.EditHint() { ContentDataPropertyName = "PageStartPublish" })
-        </div>
-    </p>
-
-    <div class="row">
-
-        @Html.Action("Full", "Category", new { currentPage = Model.CurrentPage })
-
-    </div>
-
-
-
-    <article class="main @(PageEditing.PageIsInEditMode ? "inEditMode" : "")">
-        @if (Model.CurrentPage.EasyToReadPage != null)
-        {
-            <a href="@Html.ContentLink(Model.CurrentPage.EasyToReadPage)">@Html.DisplayNameFor(m => m.CurrentPage)</a>
-        }
-
-        <header>
-
-            @if (Model.CurrentPage.TitleImage != null || PageEditing.PageIsInEditMode)
-            {
-                <figure class="title">
-
-                    @if (PageEditing.PageIsInEditMode)
-                    {
-                        <div @Html.EditAttributes(m => m.CurrentPage.TitleImage)>
-
-                            @if (Model.CurrentPage.TitleImage != null)
-                            {
-                                <img src="@Url.ContentUrl(Model.CurrentPage.TitleImage)?preset=content-title" />
-                            }
-
-                        </div>
-                    }
-                    else
-                    {
-                        <img src="@Url.ContentUrl(Model.CurrentPage.TitleImage)?preset=content-title" />
-                    }
-
-                    <figcaption>@Html.PropertyFor(m => m.CurrentPage.TitleImageCaption)</figcaption>
-
-                </figure>
-            }
-
-            <h1>@Html.PropertyFor(m => m.CurrentPage.Title)</h1>
-            <p class="article-preamble" @Html.EditAttributes(x => x.CurrentPage.Preamble)>
-                @Model.CurrentPage.Preamble
-            </p>
-        </header>
-
-        <aside class="block-area">
-            @Html.PropertyFor(m => m.CurrentPage.TopBlockArea)
-        </aside>
-
-        <section class="article-body">
-            @Html.PropertyFor(m => m.CurrentPage.Content)
-        </section>
-
-        @if ((Model.CurrentPage.RelatedEServices != null && Model.CurrentPage.RelatedEServices.Any()) ||
-                 (Model.CurrentPage.RelatedDocuments != null && Model.CurrentPage.RelatedDocuments.Any()) ||
-                 (Model.CurrentPage.RelatedLinks != null && Model.CurrentPage.RelatedLinks.Any()))
-        {
-            <aside class="related-info">
-                <h2>Relaterad information</h2>
-                <ul class="row no-bullet small-up-1 medium-up-2 link-lists">
-
-                    @if (Model.CurrentPage.RelatedEServices != null && Model.CurrentPage.RelatedEServices.Any())
-                    {
-                        <li class="column">
-                            <div>
-                                <h3>E-tjänster</h3>
-                                <ul class="link-list" @Html.EditAttributes(m => m.CurrentPage.RelatedEServices)>
-                                    @foreach (var item in Model.CurrentPage.RelatedEServices)
-                                    {
-                                        <li><a href="@item.Href" title="@item.Title" target="@item.Target">@item.Text</a></li>
-                                    }
-                                </ul>
-                            </div>
-                        </li>
-                    }
-
-                    @if (Model.CurrentPage.RelatedDocuments != null && Model.CurrentPage.RelatedDocuments.Any())
-                    {
-                        <li class="column">
-                            <div>
-                                <h3>Dokument</h3>
-                                <ul class="link-list files" @Html.EditAttributes(m => m.CurrentPage.RelatedDocuments)>
-                                    @foreach (var item in Model.CurrentPage.RelatedDocuments)
-                                    {
-                                        var fileName = Html.GetFileName(item);
-                                        <li><a href="@Url.ContentUrl(item)" title="@fileName">@fileName <span class="file-size">(@Html.GetFileSize(item))</span></a></li>
-                                    }
-                                </ul>
-                            </div>
-                        </li>
-                    }
-
-                    @if (Model.CurrentPage.RelatedLinks != null && Model.CurrentPage.RelatedLinks.Any())
-                    {
-                        <li class="column">
-                            <div>
-                                <h3>Länkar</h3>
-                                <ul class="link-list" @Html.EditAttributes(m => m.CurrentPage.RelatedLinks)>
-                                    @foreach (var item in Model.CurrentPage.RelatedLinks)
-                                    {
-                                        <li><a href="@item.Href" title="@item.Title" target="@item.Target">@item.Text</a></li>
-                                    }
-                                </ul>
-                            </div>
-                        </li>
-                    }
-
-                </ul>
-            </aside>
-        }
-
-        <aside class="block-area">
-            @Html.PropertyFor(m => m.CurrentPage.BlockArea)
-        </aside>
-
-        <footer>
-            <p>Senast uppdaterad: @Model.CurrentPage.Saved.ToString("d MMM yyyy kl. HH:mm")</p>
-            <p>Sidansvarig: Elisabeth Andersson</p>
-            <a href="#">Skriv ut sidan</a>
-            <a href="#">Dela sidan</a>
-        </footer>
-
-    </article>
-</div>
-
-<div class="column content-page right-content small-12 medium-12 large-4">
-    @{
-        Html.RenderAction("SectionMenu", "MenuKr");
-    }
-
-    <div class="block-area">
-        @Html.PropertyFor(m => m.CurrentPage.RightBlockArea)
-    </div>
-
-</div>
