+++ Kristianstad/Source/Kristianstad/Views/OrganisationalUnit/Full.cshtml
@@ -6,17 +6,58 @@
 @model OrganisationalUnitPageModel
 
 <div class="span8 clearfix">
-   
-    <div @Html.EditAttributes("PageCategory")>
-        <span>Tags: </span>
 
-        @foreach (var item in Model.Categories)
-        {
-            <span class="icon-tags">&nbsp;</span> <a href="@item.Url">@item.Title</a>
-        }
+    @if (PageEditing.PageIsInEditMode)
+    {
+        @Html.DisplayFor(x => x.CurrentPage.SourceInfo)
+    }
+
+
+    @if (Model.CurrentPage.TitleImage != null || PageEditing.PageIsInEditMode)
+    {
+        <figure class="title">
+
+            @if (PageEditing.PageIsInEditMode)
+            {
+                <div @Html.EditAttributes(m => m.CurrentPage.TitleImage)>
+
+                    @if (Model.CurrentPage.TitleImage != null)
+                    {
+                        <img src="@Url.ContentUrl(Model.CurrentPage.TitleImage)?preset=content-title" />
+                    }
+
+                </div>
+            }
+            else
+            {
+                <img src="@Url.ContentUrl(Model.CurrentPage.TitleImage)?preset=content-title" />
+            }
+
+            <figcaption>@Html.PropertyFor(m => m.CurrentPage.TitleImageCaption)</figcaption>
+
+        </figure>
+    }
+
+    <h3>Beskrivning</h3>
+    <p @Html.EditAttributes(x => x.CurrentPage.ShortDescription)>@Model.CurrentPage.ShortDescription</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.LongDescription)>@Model.CurrentPage.LongDescription</p>
+
+    <h3>Information</h3>
+    <p @Html.EditAttributes(x => x.CurrentPage.Address)>@Model.CurrentPage.Address</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.Telephone)>@Model.CurrentPage.Telephone</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.Contact)>@Model.CurrentPage.Contact</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.Email)>@Model.CurrentPage.Email</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.OrganisationalForm)>@Model.CurrentPage.OrganisationalForm</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.Latitude)>@Model.CurrentPage.Latitude</p>
+    <p @Html.EditAttributes(x => x.CurrentPage.Longitude)>@Model.CurrentPage.Longitude</p>
+
+    <div class="row">
+        <h3>Kontaktpersoner</h3>
+        @Html.PropertyFor(x => x.CurrentPage.Contacts)
     </div>
+
+    <div class="row">
+        @Html.PropertyFor(x => x.CurrentPage.Map)
+    </div>
+
 </div>
-
-<div class="span8 clearfix" @Html.EditAttributes(x => x.MainBody)>
-        @Html.DisplayFor(x => x.MainBody)
-    </div>
