--- Kristianstad/Source/Kristianstad/Views/OrganisationalUnitBlock/Index.cshtml
@@ -1,14 +1,0 @@
-﻿@model Kristianstad.ViewModels.Compare.OrganisationalUnitBlockModel
-@{ 
-    Layout = null;
-}
-
-@if (Model != null)
-{
-    @Html.PropertyFor(x => x.CurrentBlock.Body)
-}
-
-
-@{
-    Html.ViewContext.RequestContext.RouteData.Values["currentContent"] = null;
-}
