--- Kristianstad/Source/Kristianstad/Views/OrganisationalUnitsFolderPage/Index.cshtml
@@ -1,13 +1,0 @@
-﻿@using Kristianstad.Models.Pages
-
-@model EPiCore.ViewModels.Pages.PageViewModel<OrganisationalUnitsFolderPage>
-
-@{ Layout = "~/Views/Shared/_Layout.cshtml"; }
-
-<h1 @Html.EditAttributes(x => x.CurrentPage.PageName)>@Model.CurrentPage.PageName</h1>
-<p class="introduction" @Html.EditAttributes(x => x.CurrentPage.MetaDescription)>@Model.CurrentPage.MetaDescription</p>
-<div class="row">
-   @* <div class="span8 clearfix" @Html.EditAttributes(x => x.MainBody)>
-        @Html.DisplayFor(m => m.MainBody)
-    </div>*@
-</div>
