--- Kristianstad/Source/Kristianstad/Views/Shared/OrganisationalUnitBlock.cshtml
@@ -1,14 +1,0 @@
-﻿@model Kristianstad.Business.Models.Blocks.Compare.OrganisationalUnitBlock
-
-<p>
-    @Html.Label(@Model.Name)
-    <input type="text" name="Name" value="@Model.OrganisationalUnit.Name">
-</p>
-<p>
-    @if (@Model.Description != null)
-    {
-        Html.Label(@Model.Description);
-    }
-    <input type="text" name="Description" value="@Model.OrganisationalUnit.Description">
-    
-</p>
